patch e exceptions

+++ src/main/java/com/fundatec/ti20/estacionamento/service/AssinanteService.java
@@ -1,5 +1,7 @@
 package com.fundatec.ti20.estacionamento.service;
 
+import com.fundatec.ti20.estacionamento.exceptions.ConflitoException;
+import com.fundatec.ti20.estacionamento.exceptions.ObjectNotFoundException;
 import com.fundatec.ti20.estacionamento.model.Assinante;
 import com.fundatec.ti20.estacionamento.model.Conta;
 import com.fundatec.ti20.estacionamento.repository.AssinanteRepository;
@@ -20,22 +22,34 @@ public class AssinanteService {
 
 
     public Assinante findById(Integer id) {
-        return repository.findById(id).orElseThrow(() -> new RuntimeException("Assinante não encontrado"));
+        return repository.findById(id).orElseThrow(() -> new ObjectNotFoundException("Assinante"));
     }
 
     public List<Assinante> findAll() {
+        if(repository.findAll().isEmpty()){
+            throw new ObjectNotFoundException("nenhum assinante, o banco de dados está vazio");
+        }
         return repository.findAll();
     }
 
     public Assinante salvar(Assinante assinante) {
+        if (repository.findAll().contains(assinante)){
+            throw new ConflitoException("assinante");
+        }
         return repository.save(assinante);
     }
 
     public Assinante atualizar(Assinante assinante) {
+        if (repository.findAll().contains(assinante)){
+            throw new ConflitoException("assinante");
+        }
         return repository.save(assinante);
     }
 
     public void delete(Integer id) {
+        if (!repository.findAll().contains(id)){
+            throw new ObjectNotFoundException("assinante");
+        }
         repository.deleteById(id);
     }
