+++ src/main/java/com/fundatec/ti20/estacionamento/service/TarifaService.java
@@ -1,5 +1,7 @@
 package com.fundatec.ti20.estacionamento.service;
 
+import com.fundatec.ti20.estacionamento.exceptions.ConflitoException;
+import com.fundatec.ti20.estacionamento.exceptions.ObjectNotFoundException;
 import com.fundatec.ti20.estacionamento.model.Conta;
 import com.fundatec.ti20.estacionamento.model.Tarifa;
 import com.fundatec.ti20.estacionamento.model.enums.TipoTarifa;
@@ -22,22 +24,32 @@ public class TarifaService {
 
 
     public Tarifa find(TipoTarifa tipoTarifa, TipoVeiculo tipoVeiculo) {
-        return repository.findByTipoTarifaAndTipoVeiculo(tipoTarifa, tipoVeiculo).orElseThrow(() -> new RuntimeException("Tarifa não encontrada"));
+        return repository.findByTipoTarifaAndTipoVeiculo(tipoTarifa,
+                tipoVeiculo).orElseThrow(() -> new ObjectNotFoundException("Tarifa"));
     }
 
     public List<Tarifa> findAll() {
+        if(repository.findAll().isEmpty()){
+            throw new ObjectNotFoundException("nenhuma tarifa, o banco de dados está vazio");
+        }
         return repository.findAll();
     }
 
     public Tarifa findById(Integer id) {
-        return repository.findById(id).orElseThrow(() -> new RuntimeException("Tarifa não encontrada"));
+        return repository.findById(id).orElseThrow(() -> new ObjectNotFoundException("Tarifa"));
     }
 
     public Tarifa salvar(Tarifa tarifa) {
+        if (repository.findAll().contains(tarifa)){
+            throw new ConflitoException("tarifa");
+        }
         return repository.save(tarifa);
     }
 
     public void delete(Integer id) {
+        if (!repository.findAll().contains(id)){
+            throw new ObjectNotFoundException("tarifa");
+        }
         repository.deleteById(id);
     }
